# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: BBLBB API
  version: 1.0.0
  description: |
    BBLBB 稳定业务 API。字段、错误、权限和安全语义以本文件及其引用的冻结文档为准。
    生产实现必须保持 SQLite、MySQL 8 和 MariaDB 10.11 的契约一致性。
  license:
    name: Proprietary
servers:
- url: https://{host}
  variables:
    host:
      default: forum.example.com
security: []
...
paths:
...
  "/api/v1/auth/mfa/enroll":
    post:
      tags:
      - Auth
      operationId: mfaEnroll
      security:
      - sessionCookie: []
      x-permission: mfa.enroll
      summary: Begin TOTP enrollment
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: TOTP challenge (QR-code minimal data)
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
            X-Request-ID:
              "$ref": "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TotpEnrollResponse"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/mfa/confirm":
    post:
      tags:
      - Auth
      operationId: mfaConfirm
      security:
      - sessionCookie: []
      x-permission: mfa.enroll
      summary: Confirm TOTP enrollment with a 6-digit code
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MfaConfirmRequest"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/mfa/enrollment":
    delete:
      tags:
      - Auth
      operationId: mfaCancel
      security:
      - sessionCookie: []
      x-permission: mfa.enroll
      summary: Cancel a pending TOTP enrollment
      x-csrf: true
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '404':
          "$ref": "#/components/responses/NotFound"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/mfa/recovery-codes":
    post:
      tags:
      - Auth
      operationId: mfaRecoveryCodes
      security:
      - sessionCookie: []
      x-permission: mfa.recovery_codes
      summary: Generate a new set of recovery codes (shown once)
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: New recovery codes (plaintext returned exactly once)
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
            X-Request-ID:
              "$ref": "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RecoveryCodesResult"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/StepUpRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/mfa":
    delete:
      tags:
      - Auth
      operationId: mfaDisable
      security:
      - sessionCookie: []
      x-permission: mfa.disable
      summary: Disable TOTP (also invalidates all recovery codes)
      x-csrf: true
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/StepUpRequired"
        '404':
          "$ref": "#/components/responses/NotFound"
        '429':
          "$ref": "#/components/responses/RateLimited"
...
  "/api/v1/auth/re-auth":
    post:
      tags:
      - Auth
      operationId: reAuth
      security:
      - sessionCookie: []
      x-permission: mfa.reauth
      summary: Re-authenticate the current session (step-up)
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReAuthRequest"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '429':
          "$ref": "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bblbb_session
      description: Browser session cookie. State-changing requests also require X-CSRF-Token.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    RequestId:
      description: UUID v7 request correlation ID.
      schema:
        type: string
        format: uuid
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
...
  responses:
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    StepUpRequired:
      description: Recent authentication required (M02-MFA-07 step-up); re-authenticate via /api/v1/auth/re-auth
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    NotFound:
      description: Resource is missing or intentionally hidden
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    RateLimited:
      description: Request rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    GenericSuccess:
      description: Successful operation
      headers:
        Cache-Control:
          description: private, no-store
          schema:
            type: string
            example: private, no-store
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/GenericSuccess"
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
...
    GenericSuccess:
      type: object
      additionalProperties: true
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum:
          - invalid_request
          - visibility_level_exceeds_author
          - invalid_url
          - idempotency_conflict
          - version_conflict
          - unauthorized
          - invalid_credentials
          - mfa_code_invalid
          - mfa_challenge_invalid
          - mfa_confirm_invalid
          - mfa_enrollment_invalid
          - passkey_not_configured
          - passkey_unavailable
          - passkey_challenge_invalid
          - passkey_registration_invalid
          - passkey_limit_reached
          - reauth_password_invalid
          - verification_token_invalid
          - reset_token_invalid
          - validation_failed
          - invalid_current_password
          - forbidden
          - step_up_required
          - not_found
          - csrf_failed
          - origin_not_allowed
          - host_not_allowed
          - rate_limited
          - crawler_denied
          - challenge_required
          - temporarily_banned
          - feature_disabled
          - insufficient_funds
          - daily_limit_exceeded
          - checkout_interaction_invalid
          - checkout_user_mismatch
          - checkout_intent_expired
          - checkout_intent_consumed
          - offer_version_changed
          - refund_not_allowed
          - product_unavailable
          - shop_purchase_limit_exceeded
          - shop_stock_exhausted
          - entitlement_not_usable
          - presentation_slot_conflict
          - activity_already_claimed
          - activity_not_eligible
          - invalid_price_coin
          - post_not_paid
          - price_not_configured
          - self_reaction
          - cannot_follow_self
          - cannot_message_self
          - download_url_unavailable
          - ai_consent_required
          - ai_budget_exceeded
          - ai_suggestion_stale
          - invalid_storage_request
          - storage_partial_upload
          - storage_forbidden
          - storage_auth_failed
          - storage_rate_limited
          - quota_exceeded
          - storage_conflict
          - storage_state_error
          - storage_verification_failed
          - storage_network_error
          - storage_upstream_error
          - theme_invalid
          - theme_incompatible
          - theme_not_found
          - theme_conflict
          - plugin_invalid
          - plugin_incompatible
          - plugin_not_found
          - plugin_conflict
          - marketplace_disabled
          - marketplace_invalid_client
          - refund_exceeds_purchase
          - merchant_balance_insufficient
          - webhook_invalid_signature
          - bad_request
          - conflict
          - video_insecure_scheme
          - video_invalid_url
          - video_host_invalid
          - video_port_not_allowed
          - video_private_ip
          - video_signed_url
          - video_userinfo_not_allowed
          - video_fragment_not_allowed
          - video_unsupported_type
          - video_not_video_page
          - video_invalid
          - video_mime_mismatch
          - video_no_embed_permission
          - video_takedown
          - video_provider_disabled
          - video_provider_host_not_allowed
          - video_provider_ratelimited
          - video_provider_unavailable
          - video_policy_changed
          - video_policy_version_conflict
          - video_poster_attachment_invalid
          - video_resolution_expired
          - video_embed_not_found
          - video_embed_referenced
          - video_target_conflict
          - video_target_forbidden
          - video_target_not_found
          - video_version_conflict
          - video_egress_http_error
          - video_egress_private_ip
          - video_egress_timeout
          - video_egress_too_large
          - video_egress_too_many_redirects
          - video_egress_unavailable
          - video_hls_invalid
          - video_hls_depth_exceeded
          - video_hls_segment_count_exceeded
          - video_hls_duration_exceeded
          - video_hls_cross_origin_segment
          - video_hls_key_not_allowed
          - video_hls_map_not_allowed
          - video_hls_signed_uri
          - internal_error
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            type: object
            required:
            - code
            - message_key
            properties:
              field:
                type: string
              code:
                type: string
              message_key:
                type: string
...
    TotpEnrollResponse:
      type: object
      additionalProperties: false
      required:
      - otpauth_uri
      - secret_base32
      - issuer
      - account
      properties:
        otpauth_uri:
          type: string
          description: otpauth:// URI（二维码内容，M02-MFA-02 二维码所需最小数据）
          writeOnly: true
        secret_base32:
          type: string
          description: base32 编码 TOTP secret（供手工录入）
          writeOnly: true
        issuer:
          type: string
        account:
          type: string
    MfaConfirmRequest:
      type: object
      additionalProperties: false
      required:
      - code
      properties:
        code:
          type: string
          pattern: "^[0-9]{6}$"
          writeOnly: true
    ReAuthRequest:
      type: object
      additionalProperties: false
      required:
      - password
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true
    RecoveryCodesResult:
      type: object
      additionalProperties: false
      required:
      - codes
      - only_shown_once
      properties:
        codes:
          type: array
          items:
            type: string
            pattern: "^[A-Z2-7]{16}$"
          description: 明文恢复码，仅在本次响应出现一次
          writeOnly: true
        only_shown_once:
          type: boolean
          const: true
...